% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{airpassengers}
\alias{airpassengers}
\title{Air passengers}
\format{
a data frame of 144 observations containing the following variables
\describe{
\item{\code{passengers}}{number of international air passengers (in thousands).}
\item{\code{month}}{integer indicating the month of the measurement, ranging from January (\code{1}) to December (\code{12}).}
\item{\code{year}}{year of measurement}
}
}
\source{
Box, G. E. P., Jenkins, G. M. and Reinsel, G. C. (1976) \emph{Time Series Analysis, Forecasting and Control}. Third Edition. Holden-Day. Series G.
}
\usage{
airpassengers
}
\description{
Monthly totals of international airline passengers, 1949 to 1960.
}
\seealso{
\link[datasets]{AirPassengers}
}
\keyword{datasets}
