% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{auto}
\alias{auto}
\title{Automobile characteristics}
\format{
a data frame of 392 observations on the following variables
\describe{
\item{\code{mpg}}{fuel consumption (in miles per US gallon)}
\item{\code{cylinders}}{factor, the number of cylinders in the car}
\item{\code{displacement}}{engine displacement (in cubic inches)}
\item{\code{horsepower}}{engine horsepower}
\item{\code{weight}}{vehicle weight (in pounds)}
\item{\code{acceleration}}{time to accelerate from 0 to 60 mph (in seconds)}
}
}
\source{
The dataset was collected by Ernesto Ramos and David Donoho and is provided by the StatLib library at Carnegie Mellon University.
}
\usage{
auto
}
\description{
Fuel consumption and power of automobiles. The dataset was used in the 1983 American Statistical Association Exposition.
}
\references{
James, G., Witten, D., Hastie, T., and Tibshirani, R. (2013) \emph{An Introduction to Statistical Learning with applications in R}, Springer-Verlag, New York
}
\seealso{
\link[ISLR]{Auto}
}
\keyword{datasets}
