% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bixi}
\alias{bixi}
\title{Bixi rentals at Edouard-Montpetit}
\format{
a data frame of 500 observations of the following variables
\describe{
\item{\code{nusers}}{number of daily rentals (users)}
\item{\code{date}}{date of count}
\item{\code{temp}}{temperature at the nearby airport (in Celcius)}
\item{\code{relhumid}}{relative humidity (in percentage)}
\item{\code{weekday}}{week day, ranging from Sunday (\code{1}) to Saturday (\code{7})}
\item{\code{weekend}}{binary indicator for week-end, \code{0} on week-ends and \code{1} otherwise}
}
}
\source{
Bixi (bike rental), Environnement Canada
}
\usage{
bixi
}
\description{
This dataset gives the total numbers of daily rentals of Bixi bikes at the
Édouard-Montpetit station located in front of HEC. Some meteorological variables
(temperature, relative humidity) are also provided.
}
\author{
Léo Belzile
}
\keyword{datasets}
