% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bixicoll}
\alias{bixicoll}
\title{Illustration of collinearity using Bixi rentals}
\format{
a data frame of 1184 observations of the following variables
\describe{
\item{\code{lognusers}}{log of the total number of daily rentals}
\item{\code{celcius}}{temperature at the nearby airport (in Celcius)}
\item{\code{farenheit}}{temperature at the nearby airport (in Farenheit)}
\item{\code{rfarenheit}}{temperature at the nearby airport (in Farenheit), rounded to the nearest unit}
}
}
\source{
Bixi, Environnement Canada
}
\usage{
bixicoll
}
\description{
This dataset gives the total numbers of daily rentals between 2014 and 2019
of Bixi in Montreal as a function of the temperature (recorded at 16:00).
 These data are used to illustrate the effect of collinearity using rounding.
}
\author{
Léo Belzile
}
\keyword{datasets}
