% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{buchanan}
\alias{buchanan}
\title{Florida's 2000 presidential election results}
\format{
a data frame in long format of 67 observations on the following variables
\describe{
\item{\code{accidents}}{integer; daily number of accident}
\item{\code{county}}{name of county}
\item{\code{popn}}{population of the county in 1997}
\item{\code{white}}{percentage of whites in 1996}
\item{\code{black}}{percentage of blacks in 1996}
\item{\code{hisp}}{percentage of Hispanics in 1996}
\item{\code{geq65}}{percentage of the population aged 65 and above based on 1996 and 1997 population estimates}
\item{\code{highsc}}{percentage of the population with a high school degree (1990 Census data)}
\item{\code{coll}}{percentage of the population that are college graduates (1990 Census data)}
\item{\code{income}}{mean personal income in 1994}
\item{\code{buch}}{total ballots cast for Pat Buchanan (Reform)}
\item{\code{bush}}{total ballots cast for Georges W. Bush (GOP)}
\item{\code{gore}}{total ballots cast for Al Gore (Democrat)}
\item{\code{totmb}}{the total number of votes cast for the presidential election in each county, minus Buchanan votes}
}
}
\source{
Smith, R.L. (2002). \emph{A statistical assessment of Buchanan's vote in Palm Beach County}, \bold{17}(4), 441-457.
}
\usage{
buchanan
}
\description{
This database contains demographic variables and the number
of votes for the major candidates, Bush and Gore, as well
as the ballots cast for Pat Buchanan for each of the 67 counties.
}
\keyword{datasets}
