% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{college}
\alias{college}
\title{College gender inequality study}
\format{
a data frame with 397 observations of the following variables
\describe{
\item{\code{rank}}{academic rank of the professor (\code{assistant}, \code{associate} or \code{full}).}
\item{\code{field}}{categorical variable indicating whether research field is \code{applied} or \code{theoretical}.}
\item{\code{years}}{number of years since PhD}
\item{\code{sex}}{binary indicator, zero for men (\code{man}) and unity for women (\code{woman}).}
\item{\code{service}}{number of years of service in the college.}
\item{\code{salary}}{nine-month income (in thousand USD).}
}
}
\source{
Fox J. and Weisberg, S. (2019) \emph{An R Companion to Applied Regression}, Third Edition, Sage.
}
\usage{
college
}
\description{
The 2008-09 nine-month academic salary for Assistant Professors,
Associate Professors and Professors in a college in the U.S.
The data were collected as part of the on-going effort of the college's administration
to monitor salary differences between male and female faculty members.
}
\seealso{
\link[car]{Salaries}
}
\keyword{datasets}
