% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crash}
\alias{crash}
\title{Car crash on US roads}
\format{
A data frame of 40 observations (long format) containing the following five variables:
\describe{
\item{year}{factor giving the reference year, either \code{2010} or \code{2018}}
\item{region}{factor for the geographic regions defined by NHTSA}
\item{popn}{integer, total estimated population of the region based on census data}
\item{time}{factor indicating the time of the accident, \code{day} or \code{night}}
\item{ndeath}{integer, the number of death in car accidents}
}
}
\source{
\emph{National Highway Traffic Safety Administration} (NHTSA)
}
\usage{
crash
}
\description{
The National Highway Traffic Safety Administration (NHTSA) compiles
statistics about road traffic deaths in the Fatality Analysis Reporting
System. The yearly mortality counts for 2010 and 2018 are given in
crash according to whether the accident occured during daytime or
nighttime, and according to the NHTSA-defined geographic area.
}
\keyword{datasets}
