% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{distraction}
\alias{distraction}
\title{Texting while walking}
\format{
a data frame of 35 observations containing the following variables
\describe{
\item{c}{reaction time (in seconds) during a call}
\item{t}{reaction time (in seconds) while texting}
}
}
\source{
Marc Fredette
}
\usage{
distraction
}
\description{
Fake paired samples. The study undertaken at Tech3Lab
compared the reaction time (in seconds) of participants who had to
walk while texting or talking on their cell phone.
An EEG was used to record the reaction time before the candidate detected a bike
Both scenarios were randomized.
}
\keyword{datasets}
