% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gbdriving}
\alias{gbdriving}
\title{Practical driving exam results}
\format{
a data frame in long format containing the following variables, grouped by center
\describe{
 \item{pass}{number of successfull candidates (by sex)}
 \item{total}{total number of trials within the center (by sex)}
 \item{region}{factor indicating where the exam took place, either Scotland, Wales or one the nine regions of England}
 \item{sex}{factor, either \code{man} or \code{woman}}
 }
}
\source{
\url{https://www.gov.uk/government/statistical-data-sets/driving-test-statistics-drt}
}
\usage{
gbdriving
}
\description{
These records were extracted from GovUK and cleaned: they give the success rate
of practical driving license exams in the United Kingdom in 2018 in 346 driving centers.
A total of 761 750 people passed the test, out of 1 663 897 trials
}
\keyword{datasets}
