% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{insurance}
\alias{insurance}
\title{Health insurance premium}
\format{
a data frame of 1338 observations containing the following variables:
\describe{
  \item{age}{age (in years)}
  \item{sex}{sex of individual, \code{male} or \code{female}}
  \item{bmi}{body mass index (in kg/ sq.meter)}
  \item{children}{number of dependent children}
  \item{smoker}{logical, \code{yes} if the person is a smoker and \code{no} otherwise}
  \item{region}{categorical variable, the geographical location within the USA, one of \code{southwest}, \code{southeast}, \code{northwest} or \code{northeast}}
  \item{charges}{individual medical costs billed by health insurance (in USD).}
}
}
\source{
Lantz, Brett (2003), \emph{Machine Learning with R}, Packt Publishing.
}
\usage{
insurance
}
\description{
The simulated \code{insurance} data contain information about medical charges billed to 1338 Americans during 2003.
}
\keyword{datasets}
