% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{intention}
\alias{intention}
\title{Intention to buy candies based on an advertisement}
\format{
a data frame of 120 observations containing the following variables
\describe{
\item{\code{intention}}{discrete variable ranging between \code{2} and \code{14}; larger values indicate higher interest in buying the product. Specifically, the score was constructed by summing the response of two questions, both measured using a Likert scale ranging from strongly disagree (\code{1}) to strongly agree (\code{7}).}
\item{\code{fixation}}{the total duration of fixation on the ad (in seconds).}
\item{\code{emotion}}{a measure of reaction during fixation; the ratio of the probability of showing a positive emotion to the probability of showing a negative emotion.}
\item{\code{sex}}{sex of subject, either man (\code{0}) or woman (\code{1}).}
\item{\code{age}}{age (in years).}
\item{\code{revenue}}{categorical variable indicating the subject's annual income; one of (\code{1}) [0,  20k); (\code{2}) [20k,  60k); (\code{3}) 60k and above.}
\item{\code{educ}}{categorical variable indicating the highest educational achievement, either (\code{1}) high school or lower; (\code{2}) college or (\code{3}) university degree.}
\item{\code{marital}}{civil status, either single (\code{0}) or in a relationship (\code{1}).}
\item{\code{buy}}{binary variable indicating if the person bought candies.}
\item{\code{nitem}}{number of candy packs bought.}
}
}
\source{
Marc Fredette
}
\usage{
intention
}
\description{
In a study performed at Tech3Lab, subjects navigated a website that contained, among other things, an advertisement for candies. During the site navigation, an ``eye-tracker'' measured the location on the screen on which the subject's eyes were fixated.
The tracker also recorded whether the subject saw the ad and for how long it was in sight.
Additionally, facial expression analysis software (FaceReader) can be used to guess the subject's emotions when the ad was in sight.
At the end of the study, a questionnaire measured the subject's intention to buy this type of candy and socio-demographic variables.
Only the 120 subjects that had seen the ad in question are included in the data
}
\keyword{datasets}
