% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covmat.R
\name{plot.lme_rescov}
\alias{plot.lme_rescov}
\title{Plot of the covariance or correlation matrix of Y}
\usage{
\method{plot}{lme_rescov}(x, corr = FALSE, ...)
}
\arguments{
\item{x}{object of class \code{hec_rescov}}

\item{corr}{logical; if \code{TRUE}, prints the correlation matrix instead of the covariance matrix}

\item{...}{additional arguments, currently ignored}
}
\description{
Plot of the covariance or correlation matrix of Y
}
