% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{polaff}
\alias{polaff}
\title{Political affiliation of Americans}
\format{
a data frame containing the 2 by 3 contingency table
\describe{
\item{count}{total counts in sample}
\item{gender}{gender, either \code{male} or \code{female}}
\item{party}{factor for party, either Democratic Party (\code{democrat}), independent (\code{indep}) or GOP (\code{republican})}
}
}
\source{
Table 2.5 in Agresti (2007), An Introduction to Categorical Data Analysis, Wiley.
}
\usage{
polaff
}
\description{
Data drawn from the 2000 \emph{General Social Survey} containing sex and
political affiliation of surveyed individuals,
according to the party they identify the most with
}
\keyword{datasets}
