% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{renfe}
\alias{renfe}
\title{Renfe train tickets}
\format{
a data frame of 10k observations containing the following variables:
\describe{
  \item{price}{price of ticket (in euros)}
  \item{dest}{factor, binary indicator for destination (Barcelona-Madrid or Madrid-Barcelona)}
  \item{fare}{factor, categorical variable for ticket fare; one of \code{AdultoIda}, \code{Promo} or \code{Flexible}}
  \item{class}{factor, categorical variable for ticket class; one of \code{Preferente}, \code{Turista}, \code{TuristaPlus} or \code{TuristaSolo}}
  \item{type}{factor, categorical variable for train type; either \code{AVE-TGV}, \code{AVE} or \code{REXPRESS}.}
  \item{duration}{advertised duration of journey (in minutes)}
  \item{wday}{factor integer denoting week day of travel starting from Sunday(\code{1}) to Saturday (\code{7})}
}
}
\source{
Kaggle \url{https://www.kaggle.com/thegurusteam/spanish-high-speed-rail-system-ticket-pricing}
}
\usage{
renfe
}
\description{
A subset of a larger database containing information about train tickets sold by Renfe between Madrid and Barcelona.
There are three types of trains: either Spanish high speed trains Alta Velocidad Española (\code{AVE}), or
\code{AVE-TGV} which are run in partnership with SNCF to and from France (via Toulouse), or regional trains \code{REXPRESS};
only the first two labelled \code{AVE} and \code{AVE-TGV} are high speed trains.
}
\keyword{datasets}
