% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{renfe_simu}
\alias{renfe_simu}
\title{Simulation study with Renfe data}
\format{
a data frame of 1000 observations containing the following variables
\describe{
\item{meandif}{mean price difference for tickets between Madrid-Barcelona versus Barcelone-Madrid}
\item{Wstat}{Welch statistic for the two-sample t-test}
\item{cilb}{lower bound of confidence interval based on Welch statistic}
\item{ciub}{upper bound of confidence interval based on Welch statistc}
\item{pval}{p-value based on asymptotic distribution}
}
}
\source{
Léo Belzile
}
\usage{
renfe_simu
}
\description{
These are the results for a simulation study drawing sub-samples from the \code{renfe}
data, including test statistics, p-values and mean difference between samples
based on 1000 subsamples of size 100 drawn from the database.
Renfe.
}
\keyword{datasets}
