% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covmat.R
\name{rescov}
\alias{rescov}
\title{Covariance matrices for random effect models}
\source{
https://stackoverflow.com/questions/45650548/get-residual-variance-covariance-matrix-in-lme4/45655597#45655597
}
\usage{
rescov(model, ...)
}
\arguments{
\item{model}{linear mixed model fitted via \code{lme4::lme}, or an object of class \code{lmerMod}}

\item{...}{additional arguments, currently ignored}
}
\value{
an invisible list with components \code{var_y}, \code{var_raneff}, \code{var_error}
}
\description{
Covariance matrices for random effect models
}
\author{
Ben Bolker
}
