% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{waiting}
\alias{waiting}
\title{Waiting time for the Montreal metro}
\format{
a numeric vector with 62 observations
\describe{
\item{\code{waiting}}{time (in seconds) before the next metro departure}
}
}
\source{
STM
}
\usage{
waiting
}
\description{
Time (in seconds) from 17:59 until the departure of the next metro at the Universite de Montreal station
during week-days over three consecutive months
}
\keyword{datasets}
