% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{windturbine}
\alias{windturbine}
\title{Wind turbine production}
\format{
a data frame of 25 observations of the following two variables
\describe{
\item{\code{velocity}}{wind velocity (miles per hour)}
\item{\code{output}}{power generated (in DC volts)}
}
}
\source{
Joglekar et al. (1989), Lack of Fit Testing when Replicates Are Not Available, \emph{The American Statistician}, 43,(\bold{3}), 135-143.
}
\usage{
windturbine
}
\description{
The \code{windturbine} data contains measurements of electricity output of wind turbine over 25 separate fifteen minute periods. We are interested in the relation between direct output
and the average wind speed (measured in miles per hour) during the recording.
}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}. Pacific Grove, CA: Brooks/Cole, a division of Thomson Learning.
}
\seealso{
\link[BSDA]{Windmill}
}
\keyword{datasets}
